+++ src/main/java/ru/vyarus/dropwizard/guice/module/lifecycle/GuiceyLifecycleListener.java
@@ -12,6 +12,12 @@ import ru.vyarus.dropwizard.guice.module.lifecycle.event.GuiceyLifecycleEvent;
  * Listener could also implement {@link ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook}
  * in order to apply special configurations. For example, this could be some diagnostic extensions.
  * There is no other way to modify configuration in listener.
+ * <p>
+ * Listener is not registered if equal listener were already registered ({@link java.util.Set} used as
+ * listeners storage), so if you need to be sure that only one instance of some listener will be used
+ * implement {@link Object#equals(Object)}. For example, this is used to resolve case wher
+ * {@link ru.vyarus.dropwizard.guice.hook.DiagnosticHook} installed and some reports were already enabled in
+ * bundle directly: thanks to correct equals in reports, user will not see duplicate reports.
  *
  * @author Vyacheslav Rusakov
  * @see GuiceyLifecycleAdapter